Revert "fix(agents): restrict harness tool authority"
This reverts commit aacbcaacc8c9b7e1b443134c27496fbf2e229147.
