+++ extensions/codex/src/app-server/dynamic-tool-build.ts
@@ -235,7 +235,7 @@ export async function buildDynamicTools(input: DynamicToolBuildParams) {
   const agentHarnessCodingToolsFactory = injectedOpenClawCodingToolsFactory
     ? undefined
     : (dynamicToolBuildState.agentHarnessCodingToolsFactory ??
-      (await import("openclaw/plugin-sdk/agent-harness-tool-authority-runtime"))
+      (await import("openclaw/plugin-sdk/agent-harness-tool-runtime"))
         .createOpenClawCodingToolsForAgentHarness);
   const createOpenClawCodingTools =
     injectedOpenClawCodingToolsFactory ??
