+++ extensions/copilot/src/tool-bridge.ts
@@ -22,7 +22,10 @@ import {
   resolveModelAuthMode,
   sanitizeToolResult,
 } from "openclaw/plugin-sdk/agent-harness-runtime";
-import { createAgentHarnessToolSurfaceRuntime } from "openclaw/plugin-sdk/agent-harness-tool-runtime";
+import {
+  createAgentHarnessToolSurfaceRuntime,
+  createOpenClawCodingToolsForAgentHarness,
+} from "openclaw/plugin-sdk/agent-harness-tool-runtime";
 
 type CreateOpenClawCodingTools =
   (typeof import("openclaw/plugin-sdk/agent-harness"))["createOpenClawCodingTools"];
@@ -201,11 +204,8 @@ export async function createCopilotToolBridge(
   const createOpenClawCodingTools =
     input.createOpenClawCodingTools ??
     (admittedAttempt
-      ? async (options: OpenClawCodingToolsOptions) => {
-          const { createOpenClawCodingToolsForAgentHarness } =
-            await import("openclaw/plugin-sdk/agent-harness-tool-authority-runtime");
-          return createOpenClawCodingToolsForAgentHarness(admittedAttempt, options);
-        }
+      ? (options: OpenClawCodingToolsOptions) =>
+          createOpenClawCodingToolsForAgentHarness(admittedAttempt, options)
       : (await import("openclaw/plugin-sdk/agent-harness")).createOpenClawCodingTools);
 
   const toolSurfaceRuntime = createAgentHarnessToolSurfaceRuntime({
