+++ scripts/lib/plugin-sdk-entries.mjs
@@ -84,10 +84,6 @@ export const packagedPrivatePluginSdkRuntimeEntrypoints =
     productionPluginSdkEntrypointSet.has(entry),
   );
 
-const ownerRestrictedPrivatePluginSdkRuntimeEntrypointSet = new Set([
-  "agent-harness-tool-authority-runtime",
-]);
-
 /** Private entrypoints reserved for local tests and QA builds. */
 const nonProductionPrivatePluginSdkEntrypoints = privateLocalOnlyPluginSdkEntrypoints.filter(
   (entry) => !productionPluginSdkEntrypointSet.has(entry),
@@ -135,10 +131,7 @@ export function buildPluginSdkPackageExports() {
           ],
         ];
       }
-      if (
-        packagedPrivatePluginSdkRuntimeEntrypoints.includes(entry) &&
-        !ownerRestrictedPrivatePluginSdkRuntimeEntrypointSet.has(entry)
-      ) {
+      if (packagedPrivatePluginSdkRuntimeEntrypoints.includes(entry)) {
         // Official plugins ship separately but execute against the host's private runtime.
         // Their declarations stay pack-excluded by listUnpackagedPrivatePluginSdkDistArtifacts.
         return [
