+++ src/plugin-sdk/agent-harness-tool-runtime.ts
@@ -1,13 +1,21 @@
+import {
+  createOpenClawCodingToolsForAgentHarness as createCoreOpenClawCodingToolsForAgentHarness,
+  createOpenClawCodingToolsForAgentHarnessSideQuestion as createCoreOpenClawCodingToolsForAgentHarnessSideQuestion,
+} from "../agents/agent-tools-internal.js";
 /**
- * Focused runtime SDK subpath for native harness tool-surface routing.
+ * Packaged private runtime facade for official harness plugins.
  *
- * Keep tool-search and code-mode dependencies out of the lightweight harness
- * lifecycle facade used during plugin startup.
+ * This subpath is local-only and declaration-excluded. Keep tool-search and
+ * code-mode dependencies out of the lightweight startup lifecycle facade.
  */
 import {
   createAgentHarnessToolSurfaceRuntime as createCoreAgentHarnessToolSurfaceRuntime,
   type AgentHarnessToolSurfaceRuntime as CoreAgentHarnessToolSurfaceRuntime,
 } from "../agents/harness/tool-surface-bridge.js";
+import type {
+  AgentHarnessSideQuestionParams,
+  EmbeddedRunAttemptParams,
+} from "./agent-harness-runtime.js";
 
 type OpenClawCodingToolsOptions = NonNullable<
   Parameters<typeof import("./agent-harness.js").createOpenClawCodingTools>[0]
@@ -32,4 +40,26 @@ export function createAgentHarnessToolSurfaceRuntime(
   params: AgentHarnessToolSurfaceRuntimeParams,
 ): AgentHarnessToolSurfaceRuntime {
   return createCoreAgentHarnessToolSurfaceRuntime(params);
+}
+
+/**
+ * Build tools for the exact host-admitted attempt without exposing its private
+ * execution attribution to plugin code.
+ */
+export function createOpenClawCodingToolsForAgentHarness(
+  attempt: EmbeddedRunAttemptParams,
+  options?: OpenClawCodingToolsOptions,
+): ReturnType<typeof createCoreOpenClawCodingToolsForAgentHarness> {
+  return createCoreOpenClawCodingToolsForAgentHarness(attempt, options);
+}
+
+/**
+ * Build tools for the exact host-admitted side-question request without
+ * exposing its private execution attribution to plugin code.
+ */
+export function createOpenClawCodingToolsForAgentHarnessSideQuestion(
+  params: AgentHarnessSideQuestionParams,
+  options?: OpenClawCodingToolsOptions,
+): ReturnType<typeof createCoreOpenClawCodingToolsForAgentHarnessSideQuestion> {
+  return createCoreOpenClawCodingToolsForAgentHarnessSideQuestion(params, options);
 }
