+++ src/plugins/sdk-alias.test.ts
@@ -654,106 +654,6 @@ describe("plugin sdk alias helpers", () => {
     expect(shadowCodexSubpaths).toEqual(["core"]);
   });
 
-  it("aliases harness tool authority only for trusted official harness plugins", () => {
-    const fixture = createPluginSdkAliasFixture({
-      packageExports: {
-        "./plugin-sdk/core": { default: "./dist/plugin-sdk/core.js" },
-      },
-    });
-    const authoritySubpath = "agent-harness-tool-authority-runtime";
-    const sourceAuthorityPath = path.join(
-      fixture.root,
-      "src",
-      "plugin-sdk",
-      `${authoritySubpath}.ts`,
-    );
-    const distAuthorityPath = path.join(
-      fixture.root,
-      "dist",
-      "plugin-sdk",
-      `${authoritySubpath}.js`,
-    );
-    fs.writeFileSync(
-      path.join(fixture.root, "scripts", "lib", "plugin-sdk-private-local-only-subpaths.json"),
-      JSON.stringify([authoritySubpath], null, 2),
-      "utf-8",
-    );
-    fs.writeFileSync(sourceAuthorityPath, "export const authority = true;\n", "utf-8");
-    fs.writeFileSync(distAuthorityPath, "export const authority = true;\n", "utf-8");
-
-    const sourceCodexEntry = writePluginEntry(
-      fixture.root,
-      bundledPluginFile("codex", "src/index.ts"),
-    );
-    const sourceCopilotEntry = writePluginEntry(
-      fixture.root,
-      bundledPluginFile("copilot", "src/index.ts"),
-    );
-    const sourceOtherEntry = writePluginEntry(
-      fixture.root,
-      bundledPluginFile("demo", "src/index.ts"),
-    );
-    const { packageRoot: installedCodexRoot, pluginEntry: installedCodexEntry } =
-      writeInstalledPluginEntry({
-        installRoot: path.join(makeTempDir(), ".openclaw", "npm"),
-        packageName: "@openclaw/codex",
-      });
-    const { packageRoot: installedCopilotRoot, pluginEntry: installedCopilotEntry } =
-      writeInstalledPluginEntry({
-        installRoot: path.join(makeTempDir(), ".openclaw", "npm"),
-        packageName: "@openclaw/copilot",
-      });
-    const { packageRoot: installedOtherRoot, pluginEntry: installedOtherEntry } =
-      writeInstalledPluginEntry({
-        installRoot: path.join(makeTempDir(), ".openclaw", "npm"),
-        packageName: "@openclaw/demo",
-      });
-
-    const sourceCodexAliases = buildPluginLoaderAliasMap(sourceCodexEntry);
-    const sourceCopilotAliases = buildPluginLoaderAliasMap(sourceCopilotEntry);
-    const sourceOtherAliases = buildPluginLoaderAliasMap(sourceOtherEntry);
-    const installedCodexAliases = withCwd(installedCodexRoot, () =>
-      buildPluginLoaderAliasMap(
-        installedCodexEntry,
-        path.join(fixture.root, "openclaw.mjs"),
-        undefined,
-        "dist",
-      ),
-    );
-    const installedCopilotAliases = withCwd(installedCopilotRoot, () =>
-      buildPluginLoaderAliasMap(
-        installedCopilotEntry,
-        path.join(fixture.root, "openclaw.mjs"),
-        undefined,
-        "dist",
-      ),
-    );
-    const installedOtherAliases = withCwd(installedOtherRoot, () =>
-      buildPluginLoaderAliasMap(
-        installedOtherEntry,
-        path.join(fixture.root, "openclaw.mjs"),
-        undefined,
-        "dist",
-      ),
-    );
-    const specifier = `openclaw/plugin-sdk/${authoritySubpath}`;
-
-    expect(fs.realpathSync(sourceCodexAliases[specifier] ?? "")).toBe(
-      fs.realpathSync(sourceAuthorityPath),
-    );
-    expect(fs.realpathSync(sourceCopilotAliases[specifier] ?? "")).toBe(
-      fs.realpathSync(sourceAuthorityPath),
-    );
-    expect(fs.realpathSync(installedCodexAliases[specifier] ?? "")).toBe(
-      fs.realpathSync(distAuthorityPath),
-    );
-    expect(fs.realpathSync(installedCopilotAliases[specifier] ?? "")).toBe(
-      fs.realpathSync(distAuthorityPath),
-    );
-    expect(sourceOtherAliases[specifier]).toBeUndefined();
-    expect(installedOtherAliases[specifier]).toBeUndefined();
-  });
-
   it("does not reuse a non-private cached subpath list after private qa gets enabled", () => {
     const fixture = createPluginSdkAliasFixture({
       packageExports: {
